Separate parser tokens and lexer tokens (lexemes)
This separation seems to be a nice idea to increase testability for both
parser and lexer. Each of them can now be tested independently. Moreover,
none of them is now responsible for querying OpenCL devices; the parser
also is no more responsible for freeing memory, allocated for it.

This change requires introducing an intermediate entity performing
conversion from lexemes to parser tokens with necessary queries and memory
allocations returned for strings.

Signed-off-by: Pavel Kretov <firegurafiku@gmail.com>